more getComputeJobs refactor
* move to compute utils
* abstract out more loops into separate methods
* fix results display
